+++ CHANGELOG.md
@@ -12,6 +12,7 @@ If you have any data or config issues, you can run `rm -Rf ~/.awless/` to start 
 - [#55](https://github.com/wallix/awless/issues/55): Dynamically change AWS region/profile with global flags `--aws-region us-west-1` or `--aws-profile myprofile`.
 - [#73](https://github.com/wallix/awless/issues/73): `AWS_DEFAULT_REGION` env variable now loaded in `awless`. It takes precedence over `aws.region`.
 - [#73](https://github.com/wallix/awless/issues/73): `AWS_DEFAULT_PROFILE` env variable now loaded in `awless`. It takes precedence over `aws.profile`.
+- Simplification of the templating engine using decoupled compile passes
 - Better output of `awless config list` (doc per variable, etc.).
 - Config setters now provide dialogs (ex: `awless config set instance.type` or `awless config set aws.region`).
 
@@ -19,6 +20,7 @@ If you have any data or config issues, you can run `rm -Rf ~/.awless/` to start 
 
 - [#12](https://github.com/wallix/awless/issues/12): Support AWS pagination when fetching resources in AWS IAM.
 - Template parsing: allow digits in refs; allow regular chars in alias declaration
+- Template: all aliases now resolves correctly from file or CLI. Ex: `awless create instance subnet=@my-subnet`
 
 ## 0.0.16 [2017-03-01]